feat: MVVM 아키텍처를 기반으로 패키지 재구성
