+++ app/src/main/AndroidManifest.xml
@@ -13,13 +13,13 @@
         android:theme="@style/Theme.CookieTime"
         tools:targetApi="31">
         <activity
-            android:name=".ui.community.CommunityWriteActivity"
+            android:name=".presentation.community.CommunityWriteActivity"
             android:exported="false" />
         <activity
-            android:name=".ui.community.CommunityDetailActivity"
+            android:name=".presentation.community.CommunityDetailActivity"
             android:exported="false" />
         <activity
-            android:name=".ui.MainActivity"
+            android:name=".presentation.MainActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
