+++ app/src/main/java/com/ssuclass/cookietime/data/CommunityRepository.java
@@ -1,0 +1,4 @@
+package com.ssuclass.cookietime.data;
+
+public class CommunityRepository {
+}
