+++ app/src/main/res/layout/activity_community_detail.xml
@@ -5,7 +5,7 @@
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.community.CommunityDetailActivity">
+    tools:context=".presentation.community.CommunityDetailActivity">
 
     <androidx.appcompat.widget.Toolbar
         android:id="@+id/community_detail_page_toolbar"
