+++ app/src/main/res/layout/activity_community_write.xml
@@ -5,7 +5,7 @@
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.community.CommunityWriteActivity">
+    tools:context=".presentation.community.CommunityWriteActivity">
 
     <androidx.appcompat.widget.Toolbar
         android:id="@+id/community_write_toolbar_title_textview"
