+++ app/src/main/res/values/strings.xml
@@ -1,16 +1,14 @@
 <resources>
-    <!--    앱 이름-->
+    <!--앱 이름-->
     <string name="app_name">쿠키타임</string>
-    <!--    탭: 홈-->
+    <!--탭: 홈-->
     <string name="home">홈</string>
-    <!--    탭: 커뮤니티-->
+    <!--탭: 커뮤니티-->
     <string name="community">커뮤니티</string>
-    <!--    탭: 배지-->
+    <!--탭: 배지-->
     <string name="badges">배지</string>
-    <!--    탭: 마이페이지-->
+    <!--탭: 마이페이지-->
     <string name="mypage">마이페이지</string>
-
+    <!--탭: 커뮤니티 화면 네비게이션 타이틀-->
     <string name="page_title_community">전체 커뮤니티</string>
-    <!-- TODO: Remove or change this placeholder text -->
-    <string name="hello_blank_fragment">Hello blank fragment</string>
 </resources>
